Changed margins so that the spec fills more of the page.
Added 'implementation' environment and an implementation note.  Use

  \begin{implementation}
  ...
  \end{implementation}

Added roman numbering of the table of context pages and everything up
to page one.

Changed optional syntax to use [OPT] which subscripts 'opt' to
indicate an optional piece of the syntax definition.  This is the way
it is done in all the other specs that I saw (as opposed to the
brackets we used to use).  Updated syntax everywhere the syntax
listing is used.

Filled in the rest of the Primitive Types section.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7652 3a8e244f-b0f2-452b-bcba-4c88e055c3ca