Fixed a couple of bugs in arithmetic sequences:
Bounds were not snapped so

  1..10 by -2       ===>  10, 8, 6, 4, 2
  1..10 by -2 by -2 ===>  1, 5, 9

Now

  1..10 by -2       ===>  10, 8, 6, 4, 2
  1..10 by -2 by -2 ===>  2, 6, 10

Checked in a test case.

Note that

  1..10 by -2

now prints out as

  2..10 by -2

since the bounds are snapped.

Fixed a bug in the write function in which negative strides were not
written.

Added a runtime error to check for a stride of 0.

Updated a few .goods.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7672 3a8e244f-b0f2-452b-bcba-4c88e055c3ca