Added support for domain literals of the form
  [domain-specification)

where the upper bound is treated exclusively in the arithmetic domain
case so, for example, [1..10, 1..10) is equivalent to [1..9, 1..9].

A runtime error occurs if the stride of the arithmetic domain is not 1
because I didn't know what to do otherwise.

A compiler error occurs if the domain is not arithmetic.  The error
message is not very good.

Added a test case.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7693 3a8e244f-b0f2-452b-bcba-4c88e055c3ca