This checkin fixes two related bugs!   This regression is now working: arrays/bradc/stridedArrayArg2   This future is now removed: studies/blas/marybeth/saxpy1
Fixed a bug in which assigning to array arguments via whole array
assignment was being flagged as a constant assignment error.

Fixed a bug in which compound assignments, e.g., +=, are now supported
over arrays.  A cast to the type of the lhs is put in, but this
doesn't work over types that scalar promote, so the cast is ignored in
that case.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7699 3a8e244f-b0f2-452b-bcba-4c88e055c3ca