It seems that if this module has no code in it, it doesn't get used properly.  I added some bogus code for now.
I have moved the BitOps from _chpl_math to here
in my private version of the compiler, but have
not had a chance to test yet, so will test over
the weekend and check in on Monday if it works.
In doing this move, I've also cleaned some of
them up a bit/generalized them.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7788 3a8e244f-b0f2-452b-bcba-4c88e055c3ca