This check-in adds support for setting configuration variables using --varName=... in addition to the traditional -svarName=...  (I think that there was general support for this concept and that the more contentious question was whether or not to continue to support the -s flag).
This check-in also fixes the two regressions caused
by my checkin of the Memory and _chpl_machine modules.
I think my use of strings and/or a halt() routine is
what caused them to regress, because the buffer space
allocated by glom_strings before we started tracking
memory was then reallocated after we started tracking
memory.  Got around this by special-casing that pointer
in the track memory routine for now (and this should
get cleaned up when I rewrite the memory tracking code
in Chapel).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7823 3a8e244f-b0f2-452b-bcba-4c88e055c3ca