Updated implementation of + reduce (but did not update other reductions) to reflect a better semantics.
Prior, the result of a + reduce had the same type as the input type.
This was a problem for a + reduction over values of bool type because
the result of + cannot not be stored in a bool with an explicit
coercion because of a loss of data.

The new semantic is that given a reduction

  op reduce expr

the resulting type is the type of the expression

  expr op expr.

So the result of + reduce (/ true, false, true, false /) is 2.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7933 3a8e244f-b0f2-452b-bcba-4c88e055c3ca