This check-in makes the following language-level naming changes:
        "float" is now called "real"
        "imaginary" is now called "imag"
        the ".real" and ".imag" complex number
          destructurers are now ".re" and ".im"

The compiler internals have also been renamed to
use the term "real" rather than "float" in any
context where there is also an "imag" version of
the concept.  In contexts where the same value
is used for both reals and imags, it retains
the term float.  Hopefully I have been consistent
about this.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@7966 3a8e244f-b0f2-452b-bcba-4c88e055c3ca