Adding my implementations to Brad's framework.
The simple fft is in fft-candidate.

The 2d fft is in fft-candidate.  This one needs more work to be robust.
It works for size logN=8.  It should work for only even sizes of logN,
but it currently doesn't.  This implementation needs to be extended to handle
odd sizes of logN.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8019 3a8e244f-b0f2-452b-bcba-4c88e055c3ca