Removed ExprStmt and Stmt classes.  Moved subclasses of Stmt (BlockStmt, ReturnStmt, GotoStmt, CondStmt) under Expr.
Made AList contain an Expr rather than a BaseAST.  This is an
advantage since we no longer need to verify that Symbols or Types are
not in a list.  They cannot be in a list.

Moved the following fields from BaseAST to Expr: prev, next, list.
Moved related functionality to the Expr class as well.

For compiling Jacobi, this change decreases memory usage by 17% and
decreases compilation time by 11%.

To do:

  Rename and refactor parentStmt() function.  This function is
  designed to return the statement-level expression, i.e., the top
  Expr or the Expr whose parentExpr is a Stmt.  A better name is
  getStmtExpr() or simply getStmt().

  Add verification to make sure no STMT_* expressions are children of
  CallExpr.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8026 3a8e244f-b0f2-452b-bcba-4c88e055c3ca