Simplified generated code names of types and functions slightly.  Some of the type names are still rather long.
Improved copy propagation to apply to actual arguments that are in the
available set.  I was previously not doing this as I was worried about
pass by reference.  Now with references more clear, this turned out to
be easy.

Improved copy propagation to apply through casts.

These copy propagation changes decrease the LOC for the generated code
for Jacobi from 8736 to 7952 (about 9%).

Added a bunch more symbols to the "must mangle" list.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8034 3a8e244f-b0f2-452b-bcba-4c88e055c3ca