Simplfied code generation code and some file generation code.  Pruning and dead code elimination to follow.
Added to the generated code another file, called _main.c, that
contains a #include of the header file and all the module's C files.
This file is then compiled by the c compiler.  Prior, each of the C
files was compiled and preprocessed separately and each included the
header file.

This change improves compilation time by 27% when compiling Jacobi.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8081 3a8e244f-b0f2-452b-bcba-4c88e055c3ca