Reworked the implementation of && and || to use short-circuit evaluation, so the second operand is not evaluated if the first operand determines the result.  Since these can be overloaded, the semantics are defined as follows:
  expr1 && expr2  ==>  if expr1.false? then false else expr2.true?
  expr1 || expr2  ==>  if expr1.true? then true else expr2.true?

The methods false? and true? are predefined over bool type (so && and
|| are predefined over bool type).  The definitions are as follows:

  def bool.true? return this;
  def bool.false? return !this;

To overload && and ||, the user should overload the methods false? and
true? on the type(s) that the overloading is desired on.

Since conditional expressions are not yet folded, && and || are no
longer folded.

Updated array reallocate function to use & instead of && in an
expression that needs to be folded.

Added a test.

Updated the spec on logical or and logical and.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8442 3a8e244f-b0f2-452b-bcba-4c88e055c3ca