Changed implementation of fwrite and fread:
The user can now define a method called write bound to any type.  This
method must take an argument of file type.  Default write methods are
created if the user does not define one.

There is a method write also bound to file that can take any number of
arguments.  The user should always write things like

  f.write(x)

where f is a file and x is a value that they want to write.

There is also a method writeln bound to file that can take any number
of arguments.

Calls to the functions writeln and write, which can take any
arguments, call the methods writeln and write bound to stdout.

The read method is different.  The read method must always be bound to
the file type.  It's argument is inout and can be of the type that
should be read.  Multiple arguments can be passed to the read method
and this is handled, currently, by the compiler.

The advantage to having write methods bound to the file and bound to
the type being written is two-fold: 1. the locking can be handled in
the method bound to the file (thus it is good for the user to always
call this one) and 2. we dynamically dispatch based on the class being
written (I updated three test cases to this effect).

Updated many tests.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8451 3a8e244f-b0f2-452b-bcba-4c88e055c3ca