Updated the RA test in the examples directory to remove explicit iterator return types.
Added an ra-latest.chpl to the studies/hpcc
directory with the same modification, intending
hpcc06 to remain a snapshot of what we submitted
to the competition.

Added a README to the hpcc06 directory, as well as
a Makefile that builds these examples since they
are a bit more complicated than the rest of the
examples.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8469 3a8e244f-b0f2-452b-bcba-4c88e055c3ca