Prior to this check-in, we had the memory per locale hardcoded for my solitary machine.  This check-in causes the physicalMemory() routine to actually return the amount of physical memory on the locale. This is done using a new primitive, implemented in the runtime chplsys.c/h files, which uses sysconf() on most platforms, except for cygwin where it isn't defined in the same way as on other machines (the workaround is implemented in runtime/include/chplsys.h, along with a comment capturing why this workaround is necessary).
Changed the interface for physicalMemory() so that
the type used to return the value can be specified
by the user.  uint(64) is the default, but smaller
ints or floats can be desireable depending on the
locale's memory size and memory units requested.

Added a test that queries the memory in a bunch of
different ways, and added .good files for all of the
machines that we do most of our testing on.  If you
find I've made a mistake or left a machine off, feel
encouraged to make the fix yourself (hopefully verifying
by some other means that the memory returned is correct
to avoid problems like I found on cygwin).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8491 3a8e244f-b0f2-452b-bcba-4c88e055c3ca