Refactored the code that prunes the AST after function resolution (removes dead types and functions, removes method and setter tokens, ...) and added comments to this piece of code.
Improved function resolution time by almost 8% due primarily to less
AST collection traversals.

Added utility function get_string that takes an Expr* and returns a
char* if the Expr* is a string literal.  It fails if the Expr* is not
a string literal.  This complements the non-fatal get_string that
returns a bool if the Expr* is not a string literal and returns the
char* in a second argument passed by a pointer.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8562 3a8e244f-b0f2-452b-bcba-4c88e055c3ca