Fixed a bug in which prototypes in the generated code without argument lists were not declared with 'void' as the argument list.
Fixed a bug in which the few functions we put in the generated code
that are created during code generation (relating to enumerated types
and config variables) did not have prototypes.

Applied all of the runtime warning flags to the generated list of
compilation flags if the developer flag is set.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8570 3a8e244f-b0f2-452b-bcba-4c88e055c3ca