Inserted temporary variables for all formal arguments.  The temporary variables are initialized with the new special _pass function.  For all types except arrays, this calls the _copy function which is used in cases where a new variable is initialized to an expression where no type is specified.  The _copy function is equivalent to assignment in all cases.  For arrays, the _pass function simply returns the same array reference, thus allowing arrays to be passed by reference.
Since the types of formal arguments are known at the point the _pass
function is inserted, it is not strictly necessary.  However, it does
push some mechanism into the modules rather than tying policy to the
compiler, which makes the implementation more robust.

Removed three .futures.

Updated one .good as it was wrong and is now right.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8673 3a8e244f-b0f2-452b-bcba-4c88e055c3ca