Implemented inout and out to no longer incorrectly rely on pass by reference.  Arguments of in and inout are copied in.  Arguments of inout and out are copied out.
Fixed a bug in copy propagation in which it was assumed that none of
the primitives altered an argument.  The primitive fscanf does, though
this really ought to be wrapped in a primitive that just returns the
value scanned in.  Copy propagation is now disabled on all arguments
to the fscanf primitive.

Removed 6 .future files.  Fixed one of the .chpl files to not use a
variable without a type or an initializer.  Fixed one of the .good
files to assume left-to-right evaluation order so that the same
variable passed to multiple inout arguments gets its value back from
the one on the right.

Removed --formal-temps flag as this is now done.  Removed old code to
insert formal temps during normalization.

Removed handling of out arguments during normalization.  This is now
handled with the formal temp insertion.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8685 3a8e244f-b0f2-452b-bcba-4c88e055c3ca