Added two functions for traversing expressions (no symbols or types) in post-order without first collecting them all into a vector.
  Expr* getFirstExpr(expr)
  Expr* getNextExpr(expr)

Added a stylized loop for_exprs_postorder.  For example, use

  for_exprs_postorder(expr, fn->body) { ... }

to traverse the expressions in the body of a function.  The identifier
expr is a new variable of type Expr*.

Refactored internal traversal of function resolution to use this new
traversal function instead of collect_top_asts.  It has the same
semantics, but does not do the whole collection first.  This change is
in preparation for moving parameter folding into function resolution.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8879 3a8e244f-b0f2-452b-bcba-4c88e055c3ca