Removed change_types_to_values from normalization pass.
Made Symbol::isTypeVariable true for all type symbols.  This is used
on VarSymbols to signal that it is a type.  Now that type symbols are
left in the AST in the places where they were changed to variables,
this signal lets us know that it is a type.

Minor change to function resolution to evaluate the default
constructor for more types than previous.  This is necessitated
because moves are associated with less types.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@8891 3a8e244f-b0f2-452b-bcba-4c88e055c3ca